+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 This file is a running track of new features and fixes to each version of the panel released starting with `v0.0.7-alpha`.
 
+## v0.1.1-alpha
+
+### Added
+
+* Added server sorting button
+* Added user delete confirm
+* Added Copy Download Link button to backup page
+* Added registration module and REGISTRATION_ENABLED option for .env (True by default)
 
 ## v0.0.7-alpha